updated test scene to use a button to simulate a card being drawn, the script now changes day and time according to a variables of how many cards must be drawn before it changes
